Changed namespace of sql operands to be "Nevermore" to make updating code easier and because these are used in the public API of Nevermore.
