+++ source/Nevermore/SqlOperand.cs
@@ -1,0 +1,23 @@
+namespace Nevermore
+{
+    public enum ArraySqlOperand
+    {
+        In
+    }
+
+    public enum BinarySqlOperand
+    {
+        Between
+    }
+
+    public enum UnarySqlOperand
+    {
+        Equal,
+        Like,
+        GreaterThan,
+        GreaterThanOrEqual,
+        LessThan,
+        LessThanOrEqual,
+        NotEqual
+    }
+}
